requirements.txt
streamlit
google-generativeai
Pillow
SpeechRecognition
pyaudio
av
python-dotenv